+++ VVBasics/src/MutLockArray.m
@@ -24,13 +24,13 @@
 - (NSString *) description	{
 	return [NSString stringWithFormat:@"<MutLockArray: %@>",array];
 }
-+ (id) arrayWithCapacity:(int)c	{
++ (id) arrayWithCapacity:(NSUInteger)c	{
 	MutLockArray		*returnMe = [[MutLockArray alloc] initWithCapacity:c];
 	if (returnMe == nil)
 		return nil;
 	return [returnMe autorelease];
 }
-- (id) initWithCapacity:(int)c	{
+- (id) initWithCapacity:(NSUInteger)c	{
 	pthread_rwlockattr_t		attr;
 	
 	if (self = [super init])	{
@@ -179,7 +179,7 @@
 		pthread_rwlock_unlock(&arrayLock);
 	}	
 }
-- (void) insertObject:(id)o atIndex:(int)i	{
+- (void) insertObject:(id)o atIndex:(NSUInteger)i	{
 	if ((array != nil) && (o != nil) && (i<=[array count]))	{
 		//@try	{
 			[array insertObject:o atIndex:i];
@@ -189,7 +189,7 @@
 		//}
 	}
 }
-- (void) lockInsertObject:(id)o atIndex:(int)i	{
+- (void) lockInsertObject:(id)o atIndex:(NSUInteger)i	{
 	if ((array != nil) && (o != nil))	{
 		pthread_rwlock_wrlock(&arrayLock);
 			[self insertObject:o atIndex:i];
@@ -292,7 +292,7 @@
 		[self removeObject:o];
 	pthread_rwlock_unlock(&arrayLock);
 }
-- (void) removeObjectAtIndex:(int)i	{
+- (void) removeObjectAtIndex:(NSUInteger)i	{
 	if ((array!=nil) && ([array count]>0) && (i<[array count]) && (i>=0))	{
 		@try	{
 			[array removeObjectAtIndex:i];
@@ -302,7 +302,7 @@
 		}
 	}
 }
-- (void) lockRemoveObjectAtIndex:(int)i	{
+- (void) lockRemoveObjectAtIndex:(NSUInteger)i	{
 	if (array != nil)	{
 		pthread_rwlock_wrlock(&arrayLock);
 			[self removeObjectAtIndex:i];
@@ -405,7 +405,7 @@
 }
 
 
-- (id) objectAtIndex:(int)i	{
+- (id) objectAtIndex:(NSUInteger)i	{
 	if ((array != nil) && (i>=0) && (i<[array count]))	{
 		id			returnMe = nil;
 		//@try	{
@@ -418,7 +418,7 @@
 	}
 	return nil;
 }
-- (id) lockObjectAtIndex:(int)i	{
+- (id) lockObjectAtIndex:(NSUInteger)i	{
 	id		returnMe = nil;
 	pthread_rwlock_rdlock(&arrayLock);
 		returnMe = [self objectAtIndex:i];
@@ -445,8 +445,8 @@
 	
 	return returnMe;
 }
-- (int) indexOfObject:(id)o	{
-	int returnMe = -1;
+- (NSUInteger) indexOfObject:(id)o	{
+	NSUInteger returnMe = -1;
 	if ((array != nil) && (o != nil))	{
 		//@try	{
 			returnMe = [array indexOfObject:o];
@@ -457,8 +457,8 @@
 	}
 	return returnMe;	
 }
-- (int) lockIndexOfObject:(id)o	{
-	int returnMe = -1;
+- (NSUInteger) lockIndexOfObject:(id)o	{
+	NSUInteger returnMe = -1;
 	if ((array != nil) && (o != nil))	{
 		//@try	{
 			pthread_rwlock_rdlock(&arrayLock);
